requirements.txt
anyio==3.6.2
attrs==22.1.0
autopep8==2.0.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==3.4.8
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.86.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
